changed the logic of magic link
