+++ business-registry-dashboard/app/middleware/02.magic-link.global.ts
@@ -1,0 +1,13 @@
+export default defineNuxtRouteMiddleware((to) => {
+  const localePath = useLocalePath()
+
+  // Check if path contains affiliationInvitation and extract token
+  if (to.path.includes('affiliationInvitation/acceptToken/')) {
+    const token = to.path.split('acceptToken/')[1]
+    if (token) {
+      sessionStorage.setItem('affiliationToken', token)
+    }
+    window.location.href = localePath('/')
+    return
+  }
+})
